requirements.txt
asgiref==3.5.2
Django==4.1
psycopg2==2.9.3
werkzeug==2.1.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
djangorestframework==3.13.1
selenium==4.1.0
requests==2.26.0

